+++ addons/mail/__manifest__.py
@@ -141,9 +141,11 @@ For more specific needs, you may also assign custom-defined actions
             'mail/static/src/js/**/*',
             'mail/static/src/model/**/*',
             'mail/static/src/core/common/**/*',
+            'mail/static/src/core/public_web/**/*',
             'mail/static/src/core/web_portal/**/*',
             'mail/static/src/core/web/**/*',
             'mail/static/src/**/common/**/*',
+            'mail/static/src/**/public_web/**/*',
             'mail/static/src/**/web_portal/**/*',
             'mail/static/src/**/web/**/*',
             ('remove', 'mail/static/src/core/web/wysiwyg.js'),
@@ -225,7 +227,9 @@ For more specific needs, you may also assign custom-defined actions
 
             'mail/static/src/model/**/*',
             'mail/static/src/core/common/**/*',
+            'mail/static/src/core/public_web/**/*',
             'mail/static/src/**/common/**/*',
+            'mail/static/src/**/public_web/**/*',
             'mail/static/src/**/public/**/*',
             'mail/static/lib/selfie_segmentation/selfie_segmentation.js',
             ('remove', 'mail/static/src/**/*.dark.scss'),
@@ -235,8 +239,8 @@ For more specific needs, you may also assign custom-defined actions
             'mail/static/src/discuss/core/public_web/**/*',
             'mail/static/src/discuss/core/public/**/*',
             'mail/static/src/discuss/**/common/**/*',
-            'mail/static/src/discuss/**/public/**/*',
             'mail/static/src/discuss/**/public_web/**/*',
+            'mail/static/src/discuss/**/public/**/*',
             ('remove', 'mail/static/src/discuss/**/*.dark.scss'),
             ('remove', 'web/static/src/**/*.dark.scss'),
         ]
