+++ addons/mail/static/src/core/public_web/discuss.xml
