+++ addons/mail/static/src/core/web/discuss_client_action.js
@@ -1,4 +1,4 @@
-import { Discuss } from "@mail/core/common/discuss";
+import { Discuss } from "@mail/core/public_web/discuss";
 
 import { Component, onWillStart, onWillUpdateProps, useState } from "@odoo/owl";
