+++ addons/mail/static/src/discuss/public_web/discuss_patch.xml
